new(demo/geo-custom): refactor to sandbox demo
