+++ packages/vx-demo/src/components/Gallery.tsx
@@ -29,8 +29,8 @@ import Dendrograms from './tiles/Dendrograms';
 import Voronoi from './tiles/Voronoi';
 import Legends from './tiles/Legends';
 import StatsPlot from './tiles/Statsplot';
+import GeoCustom from '../docs-v2/examples/vx-geo-custom/Example';
 import GeoMercator from '../docs-v2/examples/vx-geo-mercator/Example';
-import GeoCustom from './tiles/Geo-Custom';
 import Network from './tiles/Network';
 import Streamgraph, {
   BACKGROUND as streamgraphBackgroundColor,
