+++ packages/vx-demo/src/docs-v2/examples/vx-geo-mercator/Example.tsx
@@ -19,6 +19,7 @@ interface FeatureShape {
   properties: { name: string };
 }
 
+// @ts-ignore
 const world = topojson.feature(topology, topology.objects.units) as {
   type: 'FeatureCollection';
   features: FeatureShape[];
